fix(postgres): cast OID type to int8 in GetDatabaseInfo query
The datdba column in pg_database is of OID type (type 26), but the
Go code was trying to scan it into an int64. The pgx driver's binary
protocol doesn't auto-convert OID to int64, causing E2E test failures
with error: "cannot scan oid (OID 26) in binary format into *int64"

Fix by explicitly casting d.datdba::int8 in the SQL query.

+++ internal/adapter/postgres/database.go
@@ -137,7 +137,7 @@ func (a *Adapter) GetDatabaseInfo(ctx context.Context, name string) (*types.Data
 	var info types.DatabaseInfo
 	var ownerOid int64
 	err = pool.QueryRow(ctx, `
-		SELECT d.datname, d.datdba, pg_database_size(d.datname),
+		SELECT d.datname, d.datdba::int8, pg_database_size(d.datname),
 		       pg_encoding_to_char(d.encoding), d.datcollate
 		FROM pg_database d
 		WHERE d.datname = $1`,
